requirements.txt
dask>=2023.0.0
pandas>=1.0.0
SQLAlchemy>=1.4.0
pyarrow>=10.0.0
pytest>=7.0.0
python-dotenv>=1.0.0 # Added for loading .env files
pycryptodomex>=3.20.0
psycopg2-binary==2.9.10
pyodbc==5.2.0
